flask-backend/requirements.txt
absl-py==2.1.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
awsebcli==3.20.10
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blinker==1.7.0
botocore==1.31.85
cachetools==5.3.2
cement==2.8.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.50.0
fqdn==1.5.1
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.1.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.10.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.19.0
pooch==1.8.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.17.2
PyMatting==1.1.12
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==3.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.8.5
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spotipy==2.23.0
SQLAlchemy==2.0.27
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.3.0
tifffile==2023.12.9
tinycss2==1.4.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.12
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.14.1
xgboost==2.0.3
